Fix: 3D-UNet perf sample count to 43
This sample count is not used as 3D UNet uses concatenate permutation on entire sample size. However, it's better to keep this consistent as it could confuse people.
```3d-unet.*.sample_concatenate_permutation = 1```

FWIW, due to early stopping, 3D-UNet duplicated one sample in v2.1, and it changes # of samples to be 43 from 42 in v2.0.