FIX: possible Audit TEST01 fallback accuracy check
Audit TEST01 may use fallback method when checking accuracy. In that, whole set of samples may not present in the log, and it is also possible that logs are generated for duplicate samples.
Original accuracy check script is not handling such cases properly. This fix is intended to address that shortcoming.